+++ src/links/Site.php
@@ -93,7 +93,7 @@ class Site extends Link
 
     public function getLinkText(): ?string
     {
-        if ($this->linkText) {
+        if ($this->linkText !== null && $this->linkText !== '') {
             return $this->linkText;
         }